Refactor routing structure for user and admin sections
Moved user-facing pages into a dedicated (user) directory with a shared layout, and updated admin pages to use a simplified layout. Added new admin product and sales pages, and updated navigation and dropdown menu to support role-based admin access. Cleaned up main layout and removed redundant code from shopping cart and home page.
