+++ app/(user)/favorites/page.tsx
