+++ app/(user)/layout.tsx
@@ -1,0 +1,16 @@
+import Footer from "@/components/home/Footer";
+import Navbar from "@/components/home/Navbar";
+
+export default function UserLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <Navbar />
+      <main>{children}</main>
+      <Footer />
+    </>
+  );
+}
