+++ app/(user)/page.tsx
@@ -1,0 +1,28 @@
+"use client";
+import Banner from "@/components/home/Banner";
+import FeaturedProduct from "@/components/home/FeaturedProduct";
+import HeaderSlider from "@/components/home/HeaderSlider";
+import HomeProducts from "@/components/home/HomeProducts";
+import NewsLetter from "@/components/home/NewsLetter";
+import { MessageCircle } from "lucide-react";
+
+export default function Home() {
+  return (
+    <>
+      <HeaderSlider />
+      <HomeProducts />
+      <FeaturedProduct />
+      <Banner />
+      <NewsLetter />
+      <a
+        href="https://wa.me/593980061377"
+        target="_blank"
+        rel="noopener noreferrer"
+        className="fixed bottom-4 right-4 z-50 bg-green-500 hover:bg-green-600 text-white p-3 rounded-full shadow-lg transition-colors duration-300"
+        aria-label="Chat en WhatsApp"
+      >
+        <MessageCircle size={24} />
+      </a>
+    </>
+  );
+}
