+++ app/(user)/sales/page.tsx
