+++ app/(user)/shopping-cart/page.tsx
@@ -153,15 +153,6 @@ export default function ShoppingCartPage() {
           </div>
         </div>
       )}
-      <pre>
-        {JSON.stringify(
-          {
-            cartItems,
-          },
-          null,
-          2
-        )}
-      </pre>
     </main>
   );
 }
