+++ app/admin/page.tsx
@@ -1,12 +1,10 @@
 export default function HomeAdmin() {
-    return (
-        <html lang="en" suppressHydrationWarning>
-        <body className="antialiased">
-            <main className="px-6 md:px-16 lg:px-32">
-            <h1 className="text-2xl font-bold mb-4">Admin Dashboard</h1>
-            <p>Welcome to the admin dashboard. Here you can manage your application.</p>
-            </main>
-        </body>
-        </html>
-    );
+  return (
+      <div className="px-6 md:px-16 lg:px-32">
+        <h1 className="text-2xl font-bold mb-4">Admin Dashboard</h1>
+        <p>
+          Welcome to the admin dashboard. Here you can manage your application.
+        </p>
+      </div>
+  );
 }
