+++ app/layout.tsx
@@ -4,8 +4,6 @@ import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
 import { ThemeProvider } from "@/components/theme-provider";
 import { esMX } from "@clerk/localizations";
-import Footer from "@/components/home/Footer";
-import Navbar from "@/components/home/Navbar";
 import { Toaster } from "sonner";
 
 const outfit = Outfit({ subsets: ["latin"], weight: ["300", "400", "500"] });
@@ -46,10 +44,7 @@ export default function RootLayout({
             themes={["light", "dark", "system"]}
             disableTransitionOnChange
           >
-            <header></header>
-            <Navbar />
             <main className="px-4 md:px-16 lg:px-32">{children}</main>
-            <Footer />
             <Toaster expand={false} position="top-center" richColors />
           </ThemeProvider>
         </body>
