+++ components/home/Navbar.tsx
@@ -9,7 +9,7 @@ import { ShoppingCartIconHome } from "../shared/shopping-cart";
 
 const Navbar = () => {
   return (
-    <nav className="flex items-center justify-between px-6 md:px-16 lg:px-32 py-3 border-b border-muted-foreground/20 ">
+    <nav className="flex items-center justify-between py-3 border-b border-muted-foreground/20">
       <div>
         <Link href="/">
           <LogoTheme />
